Organize MyNotes files
