+++ src/Components/CodesAndSuch/Main.jsx
@@ -1,0 +1,7 @@
+import React from 'react'
+
+export default function Main() {
+  return (
+    <div>Main</div>
+  )
+}
