+++ src/Components/MyNotes/Thumbnail.jsx
